bug in int for feature preprocessors and minor bugs in hyperparameter search space fixed

+++ autoPyTorch/pipeline/components/preprocessing/tabular_preprocessing/feature_preprocessing/Nystroem.py
@@ -14,7 +14,7 @@ import numpy as np
 import sklearn.kernel_approximation
 from sklearn.base import BaseEstimator
 
-from autoPyTorch.pipeline.components.preprocessing.tabular_preprocessing.feature_preprocessing.\
+from autoPyTorch.pipeline.components.preprocessing.tabular_preprocessing.feature_preprocessing. \
     base_feature_preprocessor import autoPyTorchFeaturePreprocessingComponent
 from autoPyTorch.utils.common import HyperparameterSearchSpace, add_hyperparameter, get_hyperparameter
 
@@ -77,9 +77,10 @@ class Nystroem(autoPyTorchFeaturePreprocessingComponent):
             else:
                 log = n_components.log
             n_components = HyperparameterSearchSpace(hyperparameter='n_components',
-                                                     value_range=(floor(int(n_components.value_range[0]) * n_features),
-                                                                  ceil(int(n_components.value_range[1]) * n_features)),
-                                                     default_value=ceil(int(n_components.default_value) * n_features),
+                                                     value_range=(
+                                                         floor(float(n_components.value_range[0]) * n_features),
+                                                         ceil(float(n_components.value_range[1]) * n_features)),
+                                                     default_value=ceil(float(n_components.default_value) * n_features),
                                                      log=log)
         else:
             n_components = HyperparameterSearchSpace(hyperparameter='n_components',
